requirements/test.txt
-r default.txt
coverage
pytest
pytest-django
pylint
